Archive old migrations
