+++ eisbuk-admin/functions/src/index.ts
@@ -6,5 +6,4 @@ export * from "./dataTriggers";
 export * from "./migrations";
 export * from "./testData";
 export * from "./testSlots";
-export * from "./oldDataModel";
 export * from "./https";
