+++ eisbuk-admin/package.json
@@ -47,6 +47,7 @@
     "start:igorice": "REACT_APP_EISBUK_SITE=igorice.web.app NODE_ENV=production node esbuild/bundler.js --mode serve",
     "start:staging": "REACT_APP_EISBUK_SITE=eisbuk.web.app NODE_ENV=production DEPLOY_STAGE=staging node esbuild/bundler.js --mode serve",
     "test:emulators": "bash -c '(trap \"kill 0\" SIGINT; ./functions/build_scripts/watch.js & firebase -c firebase-testing.json  emulators:exec --project eisbuk -- \"yarn jest --watch --runInBand src\")'",
+    "test:temp": "bash -c '(trap \"kill 0\" SIGINT; ./functions/build_scripts/watch.js & firebase -c firebase-testing.json  emulators:exec --project eisbuk -- \"yarn jest --watch --runInBand migrations\")'",
     "test:emulators:ci": "bash -c '(trap \"kill 0\" SIGINT; ./functions/build_scripts/watch.js & firebase -c firebase-testing.json  emulators:exec --project eisbuk -- \"yarn jest --ci --reporters=default --reporters=jest-junit --collect-coverage --forceExit --runInBand\")'",
     "test:quicktest": "echo Running tests with no emulators support && yarn jest --watch --env=jest-environment-jsdom-sixteen src",
     "test": "echo 'Running all tests; using emulators.' && yarn test:emulators",
