+++ eisbuk-admin/src/enums/functions.ts
@@ -8,12 +8,8 @@ export enum CloudFunction {
   CreateTestData = "createTestData",
   CreateOrganization = "createOrganization",
   CreateTestSlots = "createTestSlots",
-  CreateStaleTestData = "createStaleTestData",
 
-  MigrateSlotsToPluralCategories = "migrateSlotsToPluralCategories",
-  MigrateToNewDataModel = "migrateToNewDataModel",
   PruneSlotsByDay = "pruneSlotsByDay",
-  AddIdsToCustomers = "addIdsToCustomers",
   DeleteOrphanedBookings = "deleteOrphanedBookings",
   UnifyOffIceLabels = "unifyOffIceLabels",
 }
